+++ spec/version_spec.rb
@@ -2,6 +2,6 @@ require 'spec_helper'
 
 describe 'Leaderboard::VERSION' do
   it 'should be the correct version' do
-    Leaderboard::VERSION.should == '2.5.0'
+    Leaderboard::VERSION.should == '3.0.0.rc3'
   end
 end